Camera: UI change to reflect the selection of different camcorder profile.
Currently when camcorder profiles (High, Low, MMS or Youtube) are selected
then values for audio encoder, video encoder, video size and video duration
are set internally. Changed the UI to reflect that the above options are
not selectable (grayed out) when camcorder profiles are selected. If the
user wants to set these options, then "custom" profile should be used.

Change-Id: Iaa858bda3336f8135880b6322f1158251e6aecd0